Separate server_addon into new file
